requirements.txt
streamlit==1.45.1
python-docx==1.1.2
pandas>=2.0.0
docxtpl
streamlit-drawable-canvas
pillow 